+++ lib/generators/onesie/task_generator.rb
@@ -10,14 +10,20 @@ module Onesie
       source_root File.expand_path('./templates', __dir__)
 
       def create_task
-        template(
-          'task.rb',
-          "#{Onesie::TASKS_DIR}/#{task_version}_#{file_name}.rb",
-          class_name: class_name
-        )
+        template('task.rb', filename, class_name: class_name)
       end
 
       private
+
+      def filename
+        "#{Onesie::TASKS_DIR}/#{task_version}_#{file_name}#{task_priority}.rb"
+      end
+
+      def task_priority
+        return unless args[0]
+
+        ".#{args[0]}"
+      end
 
       def task_version
         Time.now.utc.strftime('%Y%m%d%H%M%S')